* [obd-rebase] delete chat, console command, change connection string to local db

* [obd-feature] add file sending for dishes and for updating of products and dishes (#10)

* [obd-feature] add file sending for dishes and for updating, deleting of products and dishes (#11)

* [feature] add exercises and trainers, add commands and queries to it, fix commands and queries for dishes and products, add constants (#12)

* [obd-feature] add goal (#13)

* [obd-feature] add worout and dependecies for it (#16)

* [obd-feature] add diet, fix date type of food params in product (#17)

* [obd-feature] add workout journal and fix mapping profile for extended workout (#18)

* [obd-feature] add diet journal (#19)

* Obd feature/add weight to food (#20)

* [obd-feature] add weight to food

* [obd-feature] add weight to product and init frontend with login and project page

* [obd-feature] try to add data grid

* [add] Comlete project.

+++ backend/Health.Core/Features/Products/Commands/Create/CreateProductCommand.cs
@@ -7,10 +7,10 @@ namespace Health.Core.Features.Products.Commands.Create;
 public class CreateProductCommand : IRequest<BaseResponse<long>> // вернет Id созданной сущности
 {
     public string Name { get; set; }
-    public int Calories { get; set; }
-    public int Fats { get; set; }
-    public int Proteins { get; set; }
-    public int Carbohydrates { get; set; }
+    public double Calories { get; set; }
+    public double Fats { get; set; }
+    public double Proteins { get; set; }
+    public double Carbohydrates { get; set; }
     public IFormFile? Image { get; set; }
 }
